Drop registry
