+++ toolchain/language_server/handle_initialize.cpp
@@ -2,24 +2,19 @@
 // Exceptions. See /LICENSE for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "toolchain/language_server/context.h"
-#include "toolchain/language_server/handler_registry.h"
+#include "toolchain/language_server/handle.h"
 
 namespace Carbon::LanguageServer {
 
-// Tells the client what features are supported.
-static auto HandleInitialize(
+auto HandleInitialize(
     Context& /*context*/,
     const clang::clangd::NoParams& /*client_capabilities*/,
     llvm::function_ref<void(llvm::Expected<llvm::json::Object>)> on_done)
     -> void {
   llvm::json::Object capabilities{{"documentSymbolProvider", true},
                                   {"textDocumentSync", /*Full=*/1}};
-
   llvm::json::Object reply{{"capabilities", std::move(capabilities)}};
   on_done(reply);
 }
-
-static RegisterCallHandler<HandleInitialize> register_call("initialize", "");
 
 }  // namespace Carbon::LanguageServer
