--- toolchain/language_server/handler_registry_for_notification.cpp
@@ -1,8 +1,0 @@
-// Part of the Carbon Language project, under the Apache License v2.0 with LLVM
-// Exceptions. See /LICENSE for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-#include "toolchain/language_server/handler_registry.h"
-
-// Registry instantiation needs to be one-per-file.
-LLVM_INSTANTIATE_REGISTRY(Carbon::LanguageServer::NotificationHandlerRegistry)
